+++ StripePayments/StripePayments/Source/Categories/Enums+CustomStringConvertible.swift
@@ -528,6 +528,8 @@ extension STPPaymentMethodType: CustomStringConvertible {
         case .paynow, .zip, .revolutPay, .mobilePay, .amazonPay, .alma, .konbini, .promptPay:
             // `description` is the value used when this type is converted to a string for debugging purposes, just use the display name.
             return displayName
+        case .multibanco:
+            return "multibanco"
         }
     }
 }